Move Ora.Ora methods into module and fix imports. Remove some redundant methods and change xrange to range
