+++ noora/plugins/mssql/drop/DropPlugin.py
@@ -1,6 +1,6 @@
 import os
 
-from noora.system.Ora import Ora
+from noora.system import Ora
 from noora.system.PropertyHelper import PropertyHelper
 from noora.io.File import File
